Split projects route into multiple files
