+++ routes/projects/project.js
@@ -1,0 +1,36 @@
+const express = require('express');
+const getQueryFile = require('../../utils/get-query-file');
+const getBblFeatureCollection = require('../../utils/get-bbl-feature-collection');
+
+const router = express.Router({ mergeParams: true });
+
+
+/* GET /projects/:id */
+/* Retreive a single project */
+router.get('/', async (req, res) => {
+  const { app, params } = req;
+  const { id } = params;
+
+  // import sql query templates
+  const findProjectQuery = getQueryFile('/projects/show.sql');
+
+  try {
+    const project = await app.db.one(findProjectQuery, { id });
+    project.bbl_featurecollection = await getBblFeatureCollection(project.bbls);
+
+    res.send({
+      data: {
+        type: 'projects',
+        id,
+        attributes: project,
+      },
+    });
+  } catch (e) {
+    res.status(404).send({
+      error: e.toString(),
+    });
+  }
+});
+
+
+module.exports = router;
