+++ routes/projects/tiles.js
@@ -1,0 +1,45 @@
+const express = require('express');
+const SphericalMercator = require('sphericalmercator');
+const getQueryFile = require('../../utils/get-query-file');
+
+const router = express.Router();
+const mercator = new SphericalMercator();
+
+const generateVectorTile = getQueryFile('/helpers/generate-vector-tile.sql');
+
+
+/* GET /projects/tiles/:tileId/:z/:x/:y.mvt */
+/* Retreive a vector tile by tileid */
+router.get('/:tileId/:z/:x/:y.mvt', async (req, res) => {
+  const { app, params } = req;
+
+  const {
+    tileId,
+    z,
+    x,
+    y,
+  } = params;
+
+  // retreive the projectids from the cache
+  const tileQuery = await app.tileCache.get(tileId);
+  console.log('tileQuery', tileQuery);
+  // calculate the bounding box for this tile
+  const bbox = mercator.bbox(x, y, z, false);
+
+  try {
+    const tile = await app.db.one(generateVectorTile, [...bbox, tileQuery]);
+
+    res.setHeader('Content-Type', 'application/x-protobuf');
+
+    if (tile.st_asmvt.length === 0) {
+      res.status(204);
+    }
+    res.send(tile.st_asmvt);
+  } catch (e) {
+    res.status(404).send({
+      error: e.toString(),
+    });
+  }
+});
+
+module.exports = router;
